+++ src/gtk/xattr/Ext2AttrView.cpp
@@ -1,13 +1,13 @@
 /***************************************************************************
  * ROM Properties Page shell extension. (GTK+ common)                      *
- * Ext2AttrView.c: Ext2 file system attribute viewer widget.               *
+ * Ext2AttrView.cpp: Ext2 file system attribute viewer widget.             *
  *                                                                         *
  * Copyright (c) 2017-2025 by David Korth.                                 *
  * SPDX-License-Identifier: GPL-2.0-or-later                               *
  ***************************************************************************/
 
 #include "stdafx.h"
-#include "Ext2AttrView.h"
+#include "Ext2AttrView.hpp"
 
 // Ext2 flags (also used for Ext3, Ext4, and other Linux file systems)
 #include "librpfile/xattr/ext2_flags.h"
@@ -15,11 +15,15 @@
 // Ext2AttrData
 #include "librpfile/xattr/Ext2AttrData.h"
 
+// librpfile
+using LibRpFile::XAttrReader;
+
 /* Property identifiers */
 typedef enum {
 	PROP_0,
 
 	PROP_FLAGS,
+	PROP_ZALGORITHM,
 
 	PROP_LAST
 } RpExt2AttrViewPropID;
@@ -67,12 +71,16 @@ struct _RpExt2AttrView {
 	super __parent__;
 
 	int flags;
+	XAttrReader::ZAlgorithm zAlgorithm;
 
 	// Inhibit checkbox toggling while updating.
 	gboolean inhibit_checkbox_no_toggle;
 
 	// lsattr-style attributes label
 	GtkWidget *lblLsAttr;
+
+	// Compression label
+	GtkWidget *lblCompression;
 
 	// See enum CheckboxID and checkboxInfo.
 	GtkWidget *checkBoxes[EXT2_ATTR_CHECKBOX_MAX];
@@ -103,6 +111,12 @@ rp_ext2_attr_view_class_init(RpExt2AttrViewClass *klass)
 		G_MININT, G_MAXINT, 0,
 		(GParamFlags)(G_PARAM_READWRITE | G_PARAM_STATIC_STRINGS | G_PARAM_EXPLICIT_NOTIFY));
 
+	// TODO: Enum?
+	props[PROP_ZALGORITHM] = g_param_spec_uint(
+		"zalgorithm", "zAlgorithm", "Compression algorithm",
+		0, G_MAXUINT, 0,
+		(GParamFlags)(G_PARAM_READWRITE | G_PARAM_STATIC_STRINGS | G_PARAM_EXPLICIT_NOTIFY));
+
 	// Install the properties.
 	g_object_class_install_properties(gobject_class, PROP_LAST, props);
 }
@@ -128,6 +142,14 @@ rp_ext2_attr_view_init(RpExt2AttrView *widget)
 	pango_attr_list_insert(attr_lst, pango_attr_family_new("monospace"));
 	gtk_label_set_attributes(GTK_LABEL(widget->lblLsAttr), attr_lst);
 	pango_attr_list_unref(attr_lst);
+
+	// Compression
+	widget->lblCompression = gtk_label_new("Compression:");
+	gtk_widget_set_name(widget->lblCompression, "lblCompression");
+#if GTK_CHECK_VERSION(4, 0, 0)
+	gtk_widget_set_hexpand(widget->lblCompression, true);
+	GTK_LABEL_XALIGN_RIGHT(widget->lblCompression);
+#endif /* GTK_CHECK_VERSION(4, 0, 0) */
 
 	// Checkboxes
 	int col = 0, row = 0;
@@ -179,16 +201,25 @@ rp_ext2_attr_view_init(RpExt2AttrView *widget)
 		}
 	}
 
+	// TODO: lblCompression should be right-aligned.
 #if GTK_CHECK_VERSION(4, 0, 0)
 	gtk_box_append(GTK_BOX(hboxLsAttr), lblLsAttrDesc);
 	gtk_box_append(GTK_BOX(hboxLsAttr), widget->lblLsAttr);
+	gtk_box_append(GTK_BOX(hboxLsAttr), widget->lblCompression);
 	gtk_box_append(GTK_BOX(widget), hboxLsAttr);
 	gtk_box_append(GTK_BOX(widget), gridCheckboxes);
+
+	// Don't show lblCompression by default.
+	gtk_widget_set_visible(widget->lblCompression, false);
 #else /* !GTK_CHECK_VERSION(4, 0, 0) */
 	gtk_box_pack_start(GTK_BOX(hboxLsAttr), lblLsAttrDesc, FALSE, FALSE, 0);
 	gtk_box_pack_start(GTK_BOX(hboxLsAttr), widget->lblLsAttr, FALSE, FALSE, 0);
 	gtk_box_pack_start(GTK_BOX(widget), hboxLsAttr, FALSE, FALSE, 0);
+	gtk_box_pack_end(GTK_BOX(hboxLsAttr), widget->lblCompression, FALSE, FALSE, 0);
 	gtk_widget_show_all(hboxLsAttr);
+
+	// Don't show lblCompression by default.
+	gtk_widget_hide(widget->lblCompression);
 
 	gtk_box_pack_start(GTK_BOX(widget), gridCheckboxes, FALSE, FALSE, 0);
 	gtk_widget_show_all(gridCheckboxes);
@@ -198,7 +229,7 @@ rp_ext2_attr_view_init(RpExt2AttrView *widget)
 GtkWidget*
 rp_ext2_attr_view_new(void)
 {
-	return g_object_new(RP_TYPE_EXT2_ATTR_VIEW, NULL);
+	return static_cast<GtkWidget*>(g_object_new(RP_TYPE_EXT2_ATTR_VIEW, nullptr));
 }
 
 /** Properties **/
@@ -317,6 +348,38 @@ rp_ext2_attr_view_update_flags_display(RpExt2AttrView *widget)
 	rp_ext2_attr_view_update_flags_checkboxes(widget);
 }
 
+/**
+ * Update the compression algorithm label.
+ */
+static void
+rp_ext2_attr_view_update_zAlgorithm_label(RpExt2AttrView *widget)
+{
+	const char *s_alg;
+	switch (widget->zAlgorithm) {
+		default:
+		case XAttrReader::ZAlgorithm::None:
+			// No compression...
+			s_alg = nullptr;
+			break;
+		case XAttrReader::ZAlgorithm::ZLIB:
+			s_alg = "zlib";
+			break;
+		case XAttrReader::ZAlgorithm::LZO:
+			s_alg = "lzo";
+			break;
+		case XAttrReader::ZAlgorithm::ZSTD:
+			s_alg = "zstd";
+			break;
+	}
+
+	if (s_alg) {
+		gtk_label_set_text(GTK_LABEL(widget->lblCompression), fmt::format(FSTR("Compression: {:s}"), s_alg).c_str());
+		gtk_widget_set_visible(widget->lblCompression, true);
+	} else {
+		gtk_widget_set_visible(widget->lblCompression, false);
+	}
+}
+
 /** Property accessors / mutators **/
 
 /**
@@ -360,6 +423,44 @@ rp_ext2_attr_view_clear_flags(RpExt2AttrView *widget)
 		widget->flags = 0;
 		rp_ext2_attr_view_update_flags_display(widget);
 		g_object_notify_by_pspec(G_OBJECT(widget), props[PROP_FLAGS]);
+	}
+}
+
+/**
+ * Set the current compression algorithm.
+ * @return Compression algorithm
+ */
+void
+rp_ext2_attr_view_set_zAlgorithm(RpExt2AttrView *widget, XAttrReader::ZAlgorithm zAlgorithm)
+{
+	if (widget->zAlgorithm != zAlgorithm) {
+		widget->zAlgorithm = zAlgorithm;
+		rp_ext2_attr_view_update_zAlgorithm_label(widget);
+		g_object_notify_by_pspec(G_OBJECT(widget), props[PROP_ZALGORITHM]);
+	}
+}
+
+/**
+ * Set the current compression algorithm.
+ * @param zAlgorithm Compression algorithm
+ */
+XAttrReader::ZAlgorithm
+rp_ext2_attr_view_get_zAlgorithm(RpExt2AttrView *widget)
+{
+	g_return_val_if_fail(RP_IS_EXT2_ATTR_VIEW(widget), XAttrReader::ZAlgorithm::None);
+	return widget->zAlgorithm;
+}
+
+/**
+ * Clear the current compression algorithm.
+ */
+void
+rp_ext2_attr_view_clear_zAlgorithm(RpExt2AttrView *widget)
+{
+	if (widget->zAlgorithm != XAttrReader::ZAlgorithm::None) {
+		widget->zAlgorithm = XAttrReader::ZAlgorithm::None;
+		rp_ext2_attr_view_update_zAlgorithm_label(widget);
+		g_object_notify_by_pspec(G_OBJECT(widget), props[PROP_ZALGORITHM]);
 	}
 }